move file functions to file-utils
