+++ src/utils/file-utils.ts
@@ -1,27 +1,32 @@
 import { writeFile } from 'node:fs/promises';
-
-import { format as pgformat } from '@scaleleap/pg-format';
 import { randomUUID } from 'node:crypto';
+import { performance } from 'node:perf_hooks';
+
+import JSZip from 'jszip';
+import { DuckDBConnection } from '@duckdb/node-api';
+import { format as pgformat } from '@scaleleap/pg-format';
 
 import { Dataset } from '../entities/dataset/dataset';
 import { FileImportInterface } from '../entities/dataset/file-import.interface';
 import { FileType } from '../enums/file-type';
-
 import { logger } from './logger';
 import { getFileService } from './get-file-service';
 import { asyncTmpName } from './async-tmp';
 import { duckdb } from '../services/duckdb';
 import { FACT_TABLE_NAME } from '../services/cube-builder';
 import { DataTable } from '../entities/dataset/data-table';
-import { performance } from 'node:perf_hooks';
 import { performanceReporting } from './performance-reporting';
 import { LookupTable } from '../entities/dataset/lookup-table';
 import { LookupTableExtractor } from '../extractors/lookup-table-extractor';
 import { FactTableColumn } from '../entities/dataset/fact-table-column';
 import { SUPPORTED_LOCALES, t } from '../middleware/translation';
 import { languageMatcherCaseStatement } from './lookup-table-utils';
-import { DuckDBConnection } from '@duckdb/node-api';
 import { runQueryBlock } from './run-query-block';
+import { FileImportDto } from '../dtos/file-import';
+import { FileImportType } from '../enums/file-import-type';
+import { DatasetRepository } from '../repositories/dataset';
+import { DataLakeFileEntry } from '../interfaces/datalake-file-entry';
+import { StorageService } from '../interfaces/storage-service';
 
 export const getFileImportAndSaveToDisk = async (
   dataset: Dataset,
@@ -308,3 +313,64 @@ export async function convertLookupTableToSW3Format(
   logger.debug('Converting lookup table to correct SW3 format');
   return runQueryBlock(statements);
 }
+
+export const collectFiles = async (datasetId: string): Promise<Map<string, FileImportDto>> => {
+  const files = new Map<string, FileImportDto>();
+
+  const dataset = await DatasetRepository.getById(datasetId, {
+    measure: { lookupTable: true },
+    dimensions: { lookupTable: true },
+    revisions: { dataTable: true }
+  });
+
+  if (dataset?.measure && dataset.measure.lookupTable) {
+    const fileImport = FileImportDto.fromFileImport(dataset.measure.lookupTable);
+    fileImport.type = FileImportType.Measure;
+    fileImport.parent_id = dataset.id;
+    files.set(dataset.measure.lookupTable.filename, fileImport);
+  }
+
+  dataset?.dimensions?.forEach((dimension) => {
+    if (dimension.lookupTable) {
+      const fileImport = FileImportDto.fromFileImport(dimension.lookupTable);
+      fileImport.type = FileImportType.Dimension;
+      fileImport.parent_id = dimension.id;
+      files.set(dimension.lookupTable.filename, fileImport);
+    }
+  });
+
+  dataset?.revisions?.forEach((revision) => {
+    if (revision.dataTable) {
+      const fileImport = FileImportDto.fromFileImport(revision.dataTable);
+      fileImport.type = FileImportType.DataTable;
+      fileImport.parent_id = revision.id;
+      files.set(revision.dataTable.filename, fileImport);
+    }
+  });
+
+  return files;
+};
+
+export const addDirectoryToZip = async (
+  zip: JSZip,
+  datasetFiles: Map<string, FileImportDto>,
+  directory: string,
+  fileService: StorageService
+): Promise<void> => {
+  const directoryList = await fileService.listFiles(directory);
+  for (const fileEntry of directoryList) {
+    let filename: string;
+    if ((fileEntry as DataLakeFileEntry).name) {
+      const entry = fileEntry as DataLakeFileEntry;
+      if (entry.isDirectory) {
+        await addDirectoryToZip(zip, datasetFiles, `${directory}/${entry.name}`, fileService);
+        continue;
+      }
+      filename = (fileEntry as DataLakeFileEntry).name;
+    } else {
+      filename = fileEntry as string;
+    }
+    const originalFilename = datasetFiles.get(filename)?.filename || filename;
+    zip.file(originalFilename, await fileService.loadBuffer(filename, directory));
+  }
+};
